Refactored the RabitTracker directory structure.
